% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_kallisto_index.R
\name{build_kallisto_index}
\alias{build_kallisto_index}
\title{Download ensembl transcriptome and build index for kalliso quantification}
\usage{
build_kallisto_index(indices_dir, species = "homo_sapiens", release = "94")
}
\arguments{
\item{indices_dir}{Path to directory to create indices folder in for storing
kallisto indices.}

\item{species}{The species. Default is \code{homo_sapiens}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for
AnnotationHub - needs to match with \code{\link{build_ensdb}}).}
}
\description{
This index is used for RNA seq quantification.
}
\examples{
# build kallisto index for humans
indices_dir <- getwd()
\dontrun{build_kallisto_index(indices_dir)}

}
