% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_kallisto_index.R
\name{get_pkg_version}
\alias{get_pkg_version}
\title{Get version of kallisto from system command.}
\usage{
get_pkg_version()
}
\value{
Version of kallisto.
}
\description{
Get version of kallisto from system command.
}
\examples{

get_pkg_version()

}
