% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quant_meta.R
\name{get_quant_meta}
\alias{get_quant_meta}
\title{Transform GEO metadata for quantification}
\usage{
get_quant_meta(srp_meta, data_dir)
}
\arguments{
\item{srp_meta}{data.frame of GEO metadata from
\code{\link[GEOfastq]{crawl_gsms}}}

\item{data_dir}{Directory with raw fastq.gz RNA-Seq files.}
}
\value{
data.frame needed for parameter \code{quant_meta} in
  \code{\link{run_kallisto_bulk}}
}
\description{
Transform GEO metadata for quantification
}
\examples{

# example srp_meta
srp_meta <- data.frame(run = 'SRR1',
                      gsm_name = 'GSM4875733',
                      library_layout = 'PAIRED')

# example paired fastq files
data_dir <- tempdir()
file.create(file.path(data_dir, c('SRR1_1.fastq.gz', 'SRR1_2.fastq.gz')))
quant_meta <- get_quant_meta(srp_meta, data_dir)

}
