% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs}
\alias{hs}
\title{Map from human Entrez identifier to HGNC symbol}
\format{
A data.table with 57945 rows and 2 variables:
\describe{
  \item{ENTREZID}{Entrez Identifier}
  \item{SYMBOL_9606}{HGNC symbol}
  ...
}
}
\source{
\package{crossmeta}
}
\usage{
hs
}
\description{
A dataset containing the Entrez identifiers and their corresponding HGNC symbols.
}
\keyword{datasets}
