% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{import_quants}
\alias{import_quants}
\title{Import kallisto quants}
\usage{
import_quants(data_dir, species = "Homo sapiens", release = "94")
}
\arguments{
\item{data_dir}{Directory with raw and quantified RNA-Seq files.}

\item{species}{The species. Default is \code{homo_sapiens}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for AnnotationHub -
needs to match with \code{\link{build_ensdb}}).}
}
\value{
list with items \itemize{
  \item{quants}{\code{DGEList} from \code{tximport} with argument
    \code{countsFromAbundance = 'lengthScaledTPM'}}
  \item{txi.deseq}{\code{DGEList}from \code{tximport} with argument
    \code{countsFromAbundance = 'no'}}
}
}
\description{
Import kallisto quants
}
\keyword{internal}
