% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{load_archs4_seq}
\alias{load_archs4_seq}
\title{Load ExpressionSet from ARCHS4 h5 file}
\usage{
load_archs4_seq(
  archs4_file,
  gsm_names,
  species = "Homo sapiens",
  release = "94",
  load_saved = TRUE,
  eset_path = NULL
)
}
\arguments{
\item{archs4_file}{Path to human_matrix.h5 file.}

\item{gsm_names}{Character vector of GSM names of samples to load.}

\item{species}{Character vector indicating species. Genus and species should be space separated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link{build_kallisto_index}}.}

\item{load_saved}{Should a previously saved eset be loaded? Requires \code{eset_path} to be specified.}

\item{eset_path}{Path to load saved eset from or to save eset to.}
}
\value{
ExpressionSet
}
\description{
ARCHS4 h5 files are available at https://maayanlab.cloud/archs4/download.html
}
\examples{
archs4_file <- '/path/to/human_matrix_v*.h5'
gsm_names <- c("GSM3190508", "GSM3190509", "GSM3190510", "GSM3190511")
\dontrun{eset <- load_archs4_seq(archs4_file, gsm_names)}

}
