% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{setup_fdata}
\alias{setup_fdata}
\title{Setup feature annotation data}
\usage{
setup_fdata(species = "Homo sapiens", release = "94")
}
\arguments{
\item{species}{The species. Default is \code{homo_sapiens}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for AnnotationHub -
needs to match with \code{\link{build_ensdb}}).}
}
\value{
\code{data.table} with columns \code{SYMBOL} and \code{ENTREZID_HS} corresponding to
  HGNC symbols and human entrez ids respectively.
}
\description{
Setup feature annotation data
}
