% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{construct_eset}
\alias{construct_eset}
\title{Construct expression set}
\usage{
construct_eset(quants, fdata, annot, txi.deseq = NULL)
}
\arguments{
\item{quants}{\code{DGEList} with RNA-seq counts.}

\item{fdata}{\code{data.table} returned from \code{\link{setup_fdata}}.}

\item{annot}{Character vector with ensembldb package name. e.g.
\code{'EnsDb.Hsapiens.v94'}. Returned from
\code{\link[drugseqr.data]{get_ensdb_package}}.}

\item{txi.deseq}{Optional \code{DGElist} returned by \code{import_quants}. If
specified, assays 'counts', 'abundance', and 'length' will be present in
the returned \code{ExpressionSet}.}
}
\value{
\code{ExpressionSet} with \code{quants$counts} stored in the 'exprs'
 assay, \code{quants$samples} stored in the sample metadata, and \code{fdata}
 stored in the feature data.
}
\description{
Construct expression set
}
\examples{

# generate example
y <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
row.names(y) <- paste0('gene', 1:2500)
quants <- edgeR::DGEList(counts=y)

fdata <- data.table::data.table(gene_name = row.names(y), key = 'gene_name')
annot <- drugseqr.data::get_ensdb_package('Homo sapiens', '94')

eset <- construct_eset(quants, fdata, annot)
}
