% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quant_meta.R
\name{get_quant_meta}
\alias{get_quant_meta}
\title{Transform GEO metadata for quantification}
\usage{
get_quant_meta(srp_meta, data_dir)
}
\arguments{
\item{srp_meta}{data.frame of GEO metadata from \code{\link[GEOfastq]{crawl_gsms}}}

\item{data_dir}{Directory with raw fastq.gz RNA-Seq files.}
}
\value{
data.frame needed for parameter \code{quant_meta} in \code{\link{run_kallisto_bulk}}
}
\description{
Transform GEO metadata for quantification
}
