% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{import_quants}
\alias{import_quants}
\title{Import kallisto quants}
\usage{
import_quants(data_dir, species = "Homo sapiens", release = "94")
}
\arguments{
\item{data_dir}{Directory with raw and quantified RNA-Seq files.}

\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link[drugseqr.data]{build_kallisto_index}}.}
}
\value{
\code{DGEList} with length scaled counts.
}
\description{
Import kallisto quants
}
\keyword{internal}
